+++ src/components/nearMe/types.ts
