+++ src/components/userAuth/accountCreated/DisplayResult.tsx
